refactor: dependency 변경으로 인해 폴더 위치 변경
