+++ app/api-service/src/main/java/com/cnu/spg/security/exception/LoginRequestParamterNotValidException.java
@@ -1,4 +1,4 @@
-package com.cnu.spg.user.exception;
+package com.cnu.spg.security.exception;
 
 import org.springframework.security.core.AuthenticationException;
