--- app/domain-common/src/main/java/com/cnu/spg/admin/dto/request/AdminUserRoleModifyRequest.java
@@ -1,12 +1,0 @@
-package com.cnu.spg.admin.dto.request;
-
-import lombok.Data;
-
-import java.util.Set;
-
-@Data
-public class AdminUserRoleModifyRequest {
-    private String id;
-    private String username;
-    private Set<String> roles;
-}
